feat: reliable grading + god-file split + Postgres NOTICE fix (verify.sh-green)
Makes pgmcp's quality grades trustworthy, splits the two god files, quiets
benign Postgres log spam, and lands the in-tree work-item/bug-tracker, stored
-tsvector FTS, and wfst OOV-autocorrect WIP. ./scripts/verify.sh: all gates pass.

Grading reliability (absolute thresholds, no curve; every dimension real or N/A):
- quality/report.rs: continuous DimensionScore::gpa() (score/25) replaces lossy
  letter-bucketing so a pillar's letter can't contradict its GPA; finding_density
  de-dups by file so ranking-style collectors can't saturate it to 0.
- quality/collectors/code_health.rs: complexity_hotspots/bug_prediction emit only
  absolute-threshold outliers (McCabe >10/20, lines >500/1000, bug-score >=0.4),
  not one finding per file.
- tool_engineering_scorecard.rs: complexity = real cyclomatic from function_metrics
  or honest N/A; no_god_files = absolute >2000-line bar; present-only GPA mean.
- tool_architecture_quality.rs: separation_of_concerns = N/A when topics stale.
- db/queries: topics_global_stale/graph_stale + topics_algo_signature (signature +
  freshness); tool_orient.rs uses them (real staleness, not an emptiness check).

Topic modeling: auto-derived username stopword, snake/camelCase identifier
splitting, df cutoff (in-memory + streaming paths), degeneracy-guarded signature
stamping in store_topics.

God-file split (behavior-preserving; 282 tools byte-identical; 1423 tests pass):
- src/mcp/server.rs 11,962 -> 1,461 (+ server/support.rs, mcp/params/*,
  server/handlers/* via the rmcp multi-router combine idiom).
- src/db/queries.rs 9,340 -> 79 (+ 16 #[path] per-domain submodules).
- rename the unused memory RelationRow -> MemoryRelationRow (glob-reexport clash).

Log hygiene: SET client_min_messages = warning in db/pool.rs silences benign
PostgreSQL NOTICEs surfaced at INFO ("word is too long to be indexed" from
to_tsvector skipping >2KB lexemes; "already exists, skipping" from idempotent
DDL); WARNING/ERROR still surface.

Also in this commit (concurrent in-tree WIP): first-class bug tracker (Bug kind,
Severity/BugResolution, v12 migration, work_items/bugs, triage/resolve), v13
stored-tsvector FTS, and wfst oov_autocorrect.

Docs: docs/quality/grading-reliability.md; docs/reports/report-card-2026-05-{28,29}.md.

+++ pgmcp-testing/src/fixtures/synthetic_corpus.rs
@@ -321,11 +321,11 @@ async fn seed_topic_chunks(
             let cid = insert_chunk_with_embedding(
                 pool,
                 file_id,
-                chunk_n as i32,
+                chunk_n,
                 &topic_content(topic_name, global_idx as usize),
                 &basis_with_noise(basis_idx, global_idx as usize),
-                (chunk_n as i32) * 10 + 1,
-                (chunk_n as i32) * 10 + 9,
+                chunk_n * 10 + 1,
+                chunk_n * 10 + 9,
             )
             .await;
             chunk_ids.push(cid);
@@ -443,8 +443,8 @@ async fn seed_merge_candidates(pool: &PgPool, project_id: i32) -> (i64, i64) {
                 chunk_n,
                 &topic_content("database", jit + chunk_n as usize),
                 &basis_with_noise(1, jit + chunk_n as usize),
-                (chunk_n as i32) * 5 + 1,
-                (chunk_n as i32) * 5 + 4,
+                chunk_n * 5 + 1,
+                chunk_n * 5 + 4,
             )
             .await;
         }
